Use runtime.Pinner to pin object
